Refactor playback classes into submodule
