+++ src/conq/logging/exceptions.py
@@ -1,0 +1,2 @@
+class ImageSourceUnavailableError(Exception):
+    pass
